CODE OPT PDF/Sustainable-Smart-City-Assistant-Using-IBM-Granite-LLM-main/smart-city/requirements.txt
streamlit
fastapi
uvicorn
requests
python-dotenv
pandas
scikit-learn
matplotlib
